pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2text"
version = "0.0.1"
description = "Convert scanned PDF files to searchable text via Poppler and Tesseract"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "pdf2image>=1.16",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
progress = ["tqdm>=4.66"]
rich = ["rich>=13.0"]
# Development / contributor setup
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "build>=1.0",
    "pytest-mock>=3.14",
]

[project.scripts]
pdf2text = "pdf2text.cli:main"

[tool.setuptools]
packages = ["pdf2text"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]   # E=pycodestyle errors, F=pyflakes, I=isort
ignore = ["E501"]              # handled by black or manual
target-version = "py39"

[tool.ruff.format]
quote-style = "preserve"  # preserve existing quoting
